# Conflicts:
#	libs/model/src/policies/handlers/handleReferralFeeDistributed.ts
#	libs/model/src/policies/handlers/handleReferralSet.ts
#	pnpm-lock.yaml

+++ libs/model/src/models/referral.ts
@@ -37,7 +37,7 @@ export const Referral = (
         allowNull: false,
       },
       referrer_received_eth_amount: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(78, 0),
         allowNull: false,
         defaultValue: 0,
       },
